test: rename create-common-pipe-harness
